Application structure has been reconfigured to help with code getting larger. We now have a definitions module which defines app and entry/link definitions for the holochain app. Within expressions module there now exists a file for each type of expression possible.
