+++ junto-rust/zomes/core/code/src/lib.rs
@@ -15,22 +15,19 @@ extern crate chrono;
 use multihash::Hash;
 use hdk::{
     api::DNA_ADDRESS,
-    error::{ZomeApiResult, ZomeApiError},
     holochain_core_types::{json::JsonString, hash::HashString}
 };
 
-mod user;
-mod utils;
-mod definitions;
+mod expressions;
 
 define_zome! {
     entries: [
-       user::user_definition()
+       expressions::definitions::entry_definitions::user::user_definition()
     ]
 
     genesis: || { 
         {
-            match utils::create_timestamps(HashString::encode_from_str(&DNA_ADDRESS.to_string(), Hash::SHA2256)){//Create core application timestamps "global"
+            match expressions::utils::create_timestamps(HashString::encode_from_str(&DNA_ADDRESS.to_string(), Hash::SHA2256)){//Create core application timestamps "global"
                 Ok(_) => {},
                 Err(hdk_err) => return Err(hdk_err.to_string())
             };
@@ -41,9 +38,9 @@ define_zome! {
     functions: {
         main (Public) {
             create_user: {
-                inputs: |user_data: definitions::User|,
+                inputs: |user_data: expressions::definitions::app_definitions::User|,
 				outputs: |result: JsonString|,
-				handler: user::handle_create_user
+				handler: expressions::user::handle_create_user
             }
         }
     }
